Added the beginnings of a shortcut section (I.E. the sidebar that allows for quick access to things like the Desktop, Documents, Pictures, etc.)
This was implemented with raw SDL_Texture* instead of the Texture class. This is because of the Texture classe's inability to be copied and thus worked with easily. It becomes much easier for me to understand if I just manage the memory myself to gain the ease of use that such a strategy brings (copying and the ease of use outside of the Dispaly class)