Added a copy feature
This was acomplished by having a DirObject object in the Filesystem class that stores a copy of the DirObject that is to be copyed. This DirObject can then be added to any of the Directories when the paste command is invoked along with the actual std::filesystem copy command that moves the data.