Fixed a few deleting and copying issues
For copying, the DirObject that is copied is now set to the currently selected item once it's pasted. The directory that the DirObject is pasted in is sorted again after it's pasted.

For deleting, there was an issue where the path of an object that was yanked and pasted  did not change to reflect its new position in the filesystem. This is now fixed.