New message class
Created a message class to be able to handle a bunch of diferent kinds of console prompts that I want to be able to display. Once these messages are created, they are handled in the handleMessageResponse function where different functions are called depending on the type of the Message class that was passed.