Directory rendering overhaul
Directory rendering is now much more efficient and does not use an entire CPU core when redrawing a lot in a short amount of time.

This works by doing one of two things for each direction of movement (Up or Down)
For moving down, if the selected item is not at the very bottom of the screen, the only thing that needs to be redrawn is the previous selected item and the current selected item; nothing else on the screen changes. If the selected item is at the bottom, the top most item does not need to be stored as a Texture anymore and the bottom most part of the Texture list can be added to with information of the next DirObject.

The same principle applies to moving upward, just in reverse of what was explained from downward.

Memory is now also saved by only storing the Textures that are able to be displayed on screen (with current font settings, this means there is a maximum of 17 Textures being stored and displayed for DirObjects)