Fixes #6
The problem that occured appeared in the sdl2wrapper, but did not originate there. The real issue was that an empty DirObject was attempting to be rendered, leading to exceptions being thrown. An empty DirObject was passed because of an accidental std::move statement that moved the currently selected DirObject into the FileManager class when the remove() function was called. This mean that the DirObject passed would eventually go out of scope and be deleted.

The fix was as simple as removing the std::move statement and just passing a copy to the FileManager's remove function.