SDL components Quit properly now
The display class now goes out of scope before all the SDL Quits are called for SDL, IMG, and TTF. This was not possible before because there were some unique_ptrs that had access to SDL components that were not out of scope before the Quit functions had to be called. This scoping issue is now fixed.